When delete a deployment update the `deployments_count` on program table
Close #278

+++ lambda/deployment_delete.py
@@ -1,5 +1,6 @@
 from utils import create_db_session, validate_user_access
 from decorators import validate_keys
+from models.program import Program
 from models.deployment import Deployment
 
 session = create_db_session()
@@ -16,6 +17,17 @@ def lambda_handler(event, context):
 
         validate_user_access(event, deployment)
 
+        program = session.query(Program) \
+            .filter(Program.projectcode == event['program_code']) \
+            .first()
+
+        validate_user_access(event, program)
+
+        session.query(Program) \
+            .filter(Program.projectcode == event['program_code']) \
+            .update({'deployments_count': program.deployments_count - 1})
+
+        session.flush()
         session.delete(deployment)
         session.commit()
